requirements.txt
pip==20.1.1
pyinstaller==4.9
pyserial==3.5